refactor: 디렉토리 refactor 및 미사용 이미지 제거
